Fix issue Refactor Component App
