+++ src/types/Span.tsx
@@ -1,0 +1,6 @@
+export type TypeSpan = "number" | "operation";
+export interface Span {
+	id: number;
+	type: TypeSpan;
+	value: string;
+}
